Breaking large components into smaller pieces, moving non-state based functions into utils.js files
