+++ src/components/Input/Input.scss
@@ -24,12 +24,12 @@
     font-weight: 700;
     pointer-events: none;
 
-    padding: 14px 16px;
+    padding: 14px 0;
     position: absolute;
     transition: padding 0.16s ease, font-size 0.16s ease;
 
     &.focused {
-      padding: 6px 16px;
+      padding: 6px 0;
       font-size: 10px;
     }
   }
@@ -40,8 +40,7 @@
     font-weight: 700;
     color: $white;
     background: none;
-    width: calc(100% - 32px);
-    margin: 0 16px;
+    width: 100%;
     font-family: 'Source Sans Pro', san-serif;
     border: 0;
     border-bottom: 1px solid $blue;
@@ -60,11 +59,10 @@
     border: 0;
     border-bottom: 1px solid $blue;
     padding: 20px 0 8px 0;
-    margin: 0 16px;
     font-size: 18px;
     font-weight: 700;
     color: $white;
-    width: calc(100% - 32px);
+    width: 100%;
     z-index: 1;
     transition: border 0.16s ease;
     -webkit-appearance: none;
@@ -80,8 +78,8 @@
 
 .InputSuggestions {
   position: relative;
-  width: calc(100% - 32px);
-  margin: -24px 16px 20px 16px;
+  width: 100%;
+  margin: -24px 0 20px 0;
   overflow: hidden;
   border-radius: 0 0 8px 8px;
   box-shadow: 0px 8px 8px 0 rgba(0, 0, 0, 0.32);
