+++ src/helpers/components.js
@@ -1,3 +1,4 @@
+import { useEffect, useState } from 'react'
 import { Link } from 'react-router-dom'
 
 export function ExternalLink({ url, children }) {
@@ -14,5 +15,29 @@ export function GoBack({ url, label }) {
       {'< '}
       {label}
     </Link>
+  )
+}
+
+export default function Ellipsis({ style }) {
+  const [text, setText] = useState('.')
+
+  useEffect(() => {
+    const int = window.setInterval(() => {
+      setText(text.length === 3 ? '.' : text + '.')
+    }, 400)
+    return () => window.clearInterval(int)
+  }, [text])
+
+  return (
+    <div
+      style={{
+        display: 'inline-block',
+        width: '1.1rem',
+        textAlign: 'left',
+        ...style,
+      }}
+    >
+      {text}
+    </div>
   )
 }
